Extract Hands out
